# apidefinition.json
{
    "openapi": "3.0.0",
    "info": {
      "version": "0_86",
      "title": "IPS SERN API",
      "description": "OpenAPI 3.0 specification for the IPS SERN API."
    },
    "servers": [
      {
        "url": "https://ips-d2s-uksc-medsnomed-medsno.apps.ocp1.azure.dso.digital.mod.uk",
        "description": "Production Server"
      },
      {
      "url": "http://localhost:5049",
      "description": "Local Server"
    }
  ],
...
  "paths": {
...
    "/xmpp/test-send-message": {
      "get": {
        "tags": [
          "XMPP"
        ],
        "summary": "Send a test message to XMPP group chat",
        "description": "Send a test message to the configured XMPP group chat (optionally specify ?msg=...).",
        "parameters": [
          {
            "name": "msg",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Optional message text"
          }
        ],
        "responses": {
          "200": {
            "description": "A text response indicating success.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/xmpp/xmpp-post": {
      "post": {
        "tags": [
          "XMPP"
        ],
        "summary": "Send a message to the XMPP group chat",
        "description": "Send a message to the XMPP group chat, optionally specifying a custom room.",
        "requestBody": {
          "description": "JSON body: { \"msg\": \"<message text>\", \"room\": \"<roomJid>\" } (room is optional)",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "msg": {
                    "type": "string"
                  },
                  "room": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON response with status, the target room, and the message sent.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/xmpp/xmpp-ips": {
      "post": {
        "tags": [
          "XMPP"
        ],
        "summary": "Fetch an IPS record by ID and broadcast it to the XMPP group chat",
        "description": "Fetch an IPS record by ID and broadcast it (as plain text) to the configured XMPP group chat.",
        "requestBody": {
          "description": "JSON body: { \"id\": \"<IPS record ID>\" }",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON response confirming the record was retrieved and sent.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/xmpp/xmpp-ips-private": {
      "post": {
        "tags": [
          "XMPP"
        ],
        "summary": "Fetch an IPS record by ID and send it privately to a specific occupant",
        "description": "Fetch an IPS record by ID and send it privately (type=\"chat\") to a specific occupant.",
        "requestBody": {
          "description": "JSON body: { \"id\": \"<IPS record ID>\", \"from\": \"<occupant name or user>\" }",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "from": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON response confirming the record was retrieved and sent privately.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}